chore: drop MiniCharts module, inline two helpers into WorkoutXPCard
MiniCharts.tsx exported 6 things; only clamp and seededRandom had a single caller (WorkoutXPCard). Inline them and delete the 260-line file. fitness-metrics.tsx already had its own private MiniBars copy and never imported from here.
