+++ src/components/dashboard/WorkoutXPCard.tsx
@@ -8,8 +8,26 @@ import Animated, {
 } from 'react-native-reanimated'
 import Svg, { Circle, Defs, LinearGradient, Path, Stop } from 'react-native-svg'
 
-import { clamp, seededRandom } from '@/components/charts/MiniCharts'
 import { useTheme } from '@/hooks/useThemeColor'
+
+function clamp(val: number, min: number, max: number) {
+  return Math.min(Math.max(val, min), max)
+}
+
+/** Deterministic seeded PRNG (mulberry32) — same seed → same pattern */
+function seededRandom(seed: string): () => number {
+  let h = 0
+  for (let i = 0; i < seed.length; i++) {
+    h = (Math.imul(31, h) + seed.charCodeAt(i)) | 0
+  }
+  return () => {
+    h |= 0
+    h = (h + 0x6d2b79f5) | 0
+    let t = Math.imul(h ^ (h >>> 15), 1 | h)
+    t = (t + Math.imul(t ^ (t >>> 7), 61 | t)) ^ t
+    return ((t ^ (t >>> 14)) >>> 0) / 4294967296
+  }
+}
 
 // ── Types ──────────────────────────────────────────────────────────────
